Refactor - izmjene na implementaciju prijave
